feat: add ProductList component
