+++ src/components/productList/ProductList.scss
@@ -1,0 +1,9 @@
+.product-list {
+
+	display: flex;
+	align-items: center;
+	align-content: center;
+	justify-content: center;
+	gap: 20px;
+	flex-wrap: wrap;
+}
